results_browser/requirements.txt
pixeltable
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-ag-grid==31.3.1
plotly>=5.0.0
requests>=2.25.0
gunicorn>=20.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jiwer>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
